# Conflicts:
#	src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
#	src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java

+++ src/main/java/ru/yandex/practicum/filmorate/storage/impl/UserDbStorage.java
@@ -7,8 +7,9 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.exception.IdNotFoundException;
+import ru.yandex.practicum.filmorate.model.Feed;
 import ru.yandex.practicum.filmorate.model.User;
-import ru.yandex.practicum.filmorate.storage.FilmsExtractor;
+import ru.yandex.practicum.filmorate.storage.FeedMapper;
 import ru.yandex.practicum.filmorate.storage.UserMapper;
 import ru.yandex.practicum.filmorate.storage.UserStorage;
 
@@ -21,10 +22,14 @@ import java.util.List;
 public class UserDbStorage implements UserStorage {
     private final JdbcTemplate jdbcTemplate;
     private final UserMapper userMapper;
+    private final FeedMapper feedMapper;
 
-    public UserDbStorage(JdbcTemplate jdbcTemplate, UserMapper userMapper, FilmsExtractor filmsExtractor) {
+    public UserDbStorage(JdbcTemplate jdbcTemplate,
+                         UserMapper userMapper,
+                         FeedMapper feedMapper) {
         this.jdbcTemplate = jdbcTemplate;
         this.userMapper = userMapper;
+        this.feedMapper = feedMapper;
     }
 
     @Override
@@ -35,21 +40,16 @@ public class UserDbStorage implements UserStorage {
         SimpleJdbcInsert simpleJdbcInsert = new SimpleJdbcInsert(jdbcTemplate)
                 .withTableName("users")
                 .usingGeneratedKeyColumns("id");
-
         user.setId(simpleJdbcInsert.executeAndReturnKey(user.toMap()).intValue());
         return user;
-
     }
 
     @Override
     public Collection<User> returnAllUsers() {
         String sql = "SELECT * FROM users";
-
         List<User> users = jdbcTemplate.query(
                 sql, userMapper);
-
         return users;
-
     }
 
     @Override
@@ -72,8 +72,6 @@ public class UserDbStorage implements UserStorage {
             log.error("User update failure");
             throw new IdNotFoundException("User update failure");
         }
-
-
     }
 
     @Override
@@ -92,27 +90,44 @@ public class UserDbStorage implements UserStorage {
 
     @Override
     public Collection<User> getFriendsList(int userId) {
-        String sqlQuery = "SELECT id, email, login, name, birthday " +
-                "FROM users AS u " +
-                "JOIN friends AS f ON  f.followed_user_id = u.id " +
-                "WHERE f.following_user_id = " + userId + ";";
-        return jdbcTemplate.query(sqlQuery, userMapper);
+        if (isValidUser(userId)) {
+            String sqlQuery = "SELECT id, email, login, name, birthday " +
+                    "FROM users AS u " +
+                    "JOIN friends AS f ON f.followed_user_id = u.id " +
+                    "WHERE f.following_user_id = ?";
+            return jdbcTemplate.query(sqlQuery, userMapper, userId);
+        } else {
+            throw new IdNotFoundException("User not found:" + userId);
+        }
     }
 
     @Override
-    public boolean deleteUser(int id) {
-        String sqlQuery = "delete from users where id = ?";
-        return jdbcTemplate.update(sqlQuery, id) > 0;
+    public User deleteUser(int id) {
+        if (isValidUser(id)) {
+            String sqlForFeeds = "delete from feeds where user_id in (select id from users where id = ?)";
+            jdbcTemplate.update(sqlForFeeds, id);
+            String sqlForLikes = "delete from likes where user_id in (select id from users where id = ?)";
+            jdbcTemplate.update(sqlForLikes, id);
+            String sqlForFriends = "delete from friends where following_user_id = ? or followed_user_id = ?";
+            jdbcTemplate.update(sqlForFriends, id, id);
+            String sqlForFilms = "delete from users where id = ?";
+            jdbcTemplate.update(sqlForFilms, id);
+            return null;
+        } else {
+            throw new IdNotFoundException("User not found!");
+        }
     }
 
     @Override
     public boolean addFriend(int userId1, int userId2) {
+        EventDao.eventAdd(userId2, "FRIEND", "ADD", userId1);
         String sqlQuery = "insert into friends values(?,?)";
         return jdbcTemplate.update(sqlQuery, userId1, userId2) > 0;
     }
 
     @Override
     public boolean removeFriend(int userId1, int userId2) {
+        EventDao.eventAdd(userId2, "FRIEND", "REMOVE", userId1);
         String sqlQuery = "delete from friends where following_user_id = ? AND followed_user_id = ?";
         return jdbcTemplate.update(sqlQuery, userId1, userId2) > 0;
     }
@@ -120,5 +135,11 @@ public class UserDbStorage implements UserStorage {
     @Override
     public boolean isValidUser(int id) {
         return jdbcTemplate.queryForRowSet("SELECT id FROM users WHERE id=?", id).next();
+    }
+
+    @Override
+    public List<Feed> getUserFeed(Integer id) {
+        String sqlQuery = "SELECT * FROM FEEDS f WHERE USER_ID = " + id + ";";
+        return jdbcTemplate.query(sqlQuery, feedMapper);
     }
 }
